+++ databarn/__init__.py
@@ -11,12 +11,15 @@ from .cob import Cob
 from .grain import create_grain_class as Grain
 from .barn import Barn
 from .funcs import dict_to_cob, json_to_cob
-from .decorators import one_to_many_grain, one_to_one_grain, post_init, before_assign
+from .decorators import (
+    one_to_many_grain, one_to_one_grain,
+    post_init, before_assign, after_assign)
 from .exceptions import (
-    DataBarnViolationError, DataBarnSyntaxError,
-    CobConsistencyError, StaticModelViolationError,
-    CobConstraintViolationError, GrainTypeMismatchError,
-    GrainLabelError, BarnConstraintViolationError, ValidationError)
+    DataBarnViolationError,  ValidationError,
+    DataBarnSyntaxError, CobConsistencyError,
+    StaticModelViolationError, CobConstraintViolationError,
+    GrainTypeMismatchError, GrainLabelError,
+    BarnConstraintViolationError)
 
 __version__ = "1.9.2"
 
@@ -40,4 +43,5 @@ __all__ = [
     "one_to_many_grain",
     "one_to_one_grain",
     "before_assign",
+    "after_assign",
 ]