+++ databarn/constants.py
@@ -8,3 +8,4 @@ ABSENT = Sentinel("ABSENT")  # Sentinel for absent value (unset or deleted).
 RESERVED_ATTR_NAME = '__dna__'
 POST_INIT_ATTR_NAME = "__databarn_post_init__"
 BEFORE_ASSIGN_ATTR_NAME = "__databarn_before_assign__"
+AFTER_ASSIGN_ATTR_NAME = "__databarn_after_assign__"