+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ## Changed
 - Created the exception `ValidationError`.
 - Added the `@before_assign` decorator for registering pre-assignment hooks.
+- Added the `@after_assign` decorator for registering post-assignment validation hooks.
 
 # 1.9.2